Refactor : rename directory for contants.js to Constants
